ref: Migrate `SentryBinaryImageCache` to Swift (#5995)
* ref: Migrate `SentryBinaryImageCache` to Swift

* Fix compile on xrOS

* Move headers to include folder

* Update sdk json

* Use `SentryDependencyContainerSwiftHelper` for the dispatchQueueWrapper

* Update SDK api json

* Revert to use binary image callbacks in C

* Move `SentryBinaryImageCacheCallbacks` to include folder

* Fix duplicate files

* Use `@_spi(Private)` and `final` for

* Refactor `imageByAddress` method to simplify index check and update return value in `SentryBinaryImageCache`

* Fix PR review comments

* Revert to use `self.cache?.insert`

* Revert unmutated variable to use let

+++ Sources/Sentry/SentryCrashStackEntryMapper.m
@@ -1,5 +1,4 @@
 #import "SentryCrashStackEntryMapper.h"
-#import "SentryBinaryImageCache.h"
 #import "SentryDependencyContainer.h"
 #import "SentryFormatter.h"
 #import "SentryFrame.h"
@@ -42,7 +41,7 @@ NS_ASSUME_NONNULL_BEGIN
     // we get image from the cache.
     if (stackEntry.imageAddress == 0 && stackEntry.imageName == NULL) {
         SentryBinaryImageInfo *info = [SentryDependencyContainer.sharedInstance.binaryImageCache
-            imageByAddress:stackEntry.address];
+            imageByAddress:(uint64_t)stackEntry.address];
 
         frame.imageAddress = sentry_formatHexAddressUInt64(info.address);
         frame.package = info.name;
